Type Fixes (#366)
* Add `javaType` to `PGAnyType`

Allows easy retrieval of the default class for known and resolved types.

* Fix field type in ResultField

* Add procs for `macaddr8`

* Force all text/binary decoders to have same defaultClass

* Fix network test to work on pre-10.0 servers

+++ driver/src/main/java/com/impossibl/postgres/protocol/ResultField.java
@@ -34,7 +34,7 @@ public class ResultField implements FieldFormatRef {
 
   private String name;
   private int relationId;
-  private int relationAttributeNumber;
+  private short relationAttributeNumber;
   private TypeRef typeRef;
   private short typeLength;
   private int typeModifier;
@@ -68,7 +68,7 @@ public class ResultField implements FieldFormatRef {
     return relationId;
   }
 
-  public int getRelationAttributeNumber() {
+  public short getRelationAttributeNumber() {
     return relationAttributeNumber;
   }
